--- src/main/java/roomescape/auth/application/AuthorizationException.java
@@ -1,12 +1,0 @@
-package roomescape.auth.application;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class AuthorizationException extends RuntimeException {
-
-    public AuthorizationException(String message) {
-        super(message);
-    }
-}
