+++ src/main/java/roomescape/controller/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto.request;
+package roomescape.controller.dto.request;
 
 public record LoginRequest(String email, String password) {
 }
