+++ src/main/java/roomescape/controller/view/LoginPageController.java
@@ -1,4 +1,4 @@
-package roomescape.auth.ui;
+package roomescape.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
