+++ src/main/java/roomescape/exception/custom/AuthorizationException.java
@@ -1,0 +1,8 @@
+package roomescape.exception.custom;
+
+public class AuthorizationException extends RuntimeException {
+
+    public AuthorizationException(String message) {
+        super(message);
+    }
+}
